pi/api/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
uuid7
python-dotenv